fix: ModelDto now has decimal wheel size

+++ ams-desk-cs-backend/Models/Dtos/ModelDto.cs
@@ -19,7 +19,7 @@ public class ModelDto
     public short FrameSize { get; set; }
     [Required]
     [Range(10, 30, ErrorMessage = "Niepoprawny rozmiar koła")]
-    public short WheelSize { get; set; }
+    public decimal WheelSize { get; set; }
     [Required]
     public bool IsWoman { get; set; }
     [Required]
